requirements.txt
openai
google-genai

requests
pyyaml
networkx
matplotlib
tqdm
